Rebase from 'upstream'

+++ qwt/src/qwt_matrix_raster_data.cpp
@@ -8,8 +8,11 @@
  *****************************************************************************/
 
 #include "qwt_matrix_raster_data.h"
+#include "qwt_interval.h"
+
+#include <qvector.h>
 #include <qnumeric.h>
-#include <qmath.h>
+#include <qrect.h>
 
 class QwtMatrixRasterData::PrivateData
 {
@@ -82,10 +85,10 @@ QwtMatrixRasterData::ResampleMode QwtMatrixRasterData::resampleMode() const
 
    \param axis X, Y or Z axis
    \param interval Interval
-   
+
    \sa QwtRasterData::interval(), setValueMatrix()
 */
-void QwtMatrixRasterData::setInterval( 
+void QwtMatrixRasterData::setInterval(
     Qt::Axis axis, const QwtInterval &interval )
 {
     if ( axis >= 0 && axis <= 2 )
@@ -112,7 +115,7 @@ QwtInterval QwtMatrixRasterData::interval( Qt::Axis axis ) const
 
    The positions of the values are calculated by dividing
    the bounding rectangle of the X/Y intervals into equidistant
-   rectangles ( pixels ). Each value corresponds to the center of 
+   rectangles ( pixels ). Each value corresponds to the center of
    a pixel.
 
    \param values Vector of values
@@ -120,7 +123,7 @@ QwtInterval QwtMatrixRasterData::interval( Qt::Axis axis ) const
 
    \sa valueMatrix(), numColumns(), numRows(), setInterval()()
 */
-void QwtMatrixRasterData::setValueMatrix( 
+void QwtMatrixRasterData::setValueMatrix(
     const QVector<double> &values, int numColumns )
 {
     d_data->values = values;
@@ -177,17 +180,17 @@ int QwtMatrixRasterData::numRows() const
 /*!
    \brief Calculate the pixel hint
 
-   pixelHint() returns the geometry of a pixel, that can be used 
+   pixelHint() returns the geometry of a pixel, that can be used
    to calculate the resolution and alignment of the plot item, that is
-   representing the data. 
+   representing the data.
 
    - NearestNeighbour\n
-     pixelHint() returns the surrounding pixel of the top left value 
+     pixelHint() returns the surrounding pixel of the top left value
      in the matrix.
 
    - BilinearInterpolation\n
      Returns an empty rectangle recommending
-     to render in target device ( f.e. screen ) resolution. 
+     to render in target device ( f.e. screen ) resolution.
 
    \param area Requested area, ignored
    \return Calculated hint
@@ -242,12 +245,12 @@ double QwtMatrixRasterData::value( double x, double y ) const
 
             if ( col1 < 0 )
                 col1 = col2;
-            else if ( col2 >= static_cast<int>( d_data->numColumns ) )
+            else if ( col2 >= d_data->numColumns )
                 col2 = col1;
 
             if ( row1 < 0 )
                 row1 = row2;
-            else if ( row2 >= static_cast<int>( d_data->numRows ) )
+            else if ( row2 >= d_data->numRows )
                 row2 = row1;
 
             const double v11 = d_data->value( row1, col1 );
@@ -255,11 +258,11 @@ double QwtMatrixRasterData::value( double x, double y ) const
             const double v12 = d_data->value( row2, col1 );
             const double v22 = d_data->value( row2, col2 );
 
-            const double x2 = xInterval.minValue() + 
+            const double x2 = xInterval.minValue() +
                 ( col2 + 0.5 ) * d_data->dx;
-            const double y2 = yInterval.minValue() + 
+            const double y2 = yInterval.minValue() +
                 ( row2 + 0.5 ) * d_data->dy;
-                
+
             const double rx = ( x2 - x ) / d_data->dx;
             const double ry = ( y2 - y ) / d_data->dy;
